use long instead of int in some of the count methods, use base methods for pageSize/number and order for criteria and queries, cleanup

+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/common/IAnnotationDao.java
@@ -1,6 +1,6 @@
 /**
 * Copyright (C) 2008 EDIT
-* European Distributed Institute of Taxonomy 
+* European Distributed Institute of Taxonomy
 * http://www.e-taxonomy.eu
 */
 
@@ -20,19 +20,19 @@ import eu.etaxonomy.cdm.persistence.query.OrderHint;
  * @since 12.09.2008
  */
 public interface IAnnotationDao extends ILanguageStringBaseDao<Annotation> {
-	
+
 	/**
 	 * Returns a count of Annotations, optionally filtered by commentator and status
-	 * 
+	 *
 	 * @param commentator The person who made the annotation (null to count all annotations, regardless of who made the comment)
 	 * @param status The status of the annotations (null to count all annotations regardless of status)
 	 * @return a count of Annotation instances
 	 */
-    public int count(Person commentator, MarkerType status);
-	
+    public long count(Person commentator, MarkerType status);
+
     /**
 	 * Returns a List of Annotations, optionally filtered by commentator and status
-	 * 
+	 *
 	 * @param commentator The person who made the annotation (null to list all annotations, regardless of who made the comment)
 	 * @param status The status of the annotations (null to return annotations regardless of status)
 	 * @param pageSize The maximum number of annotations returned (can be null for all annotations)
@@ -42,19 +42,19 @@ public interface IAnnotationDao extends ILanguageStringBaseDao<Annotation> {
 	 * @return a List of Annotation instances
 	 */
 	public List<Annotation> list(Person commentator, MarkerType status, Integer pageSize, Integer pageNumber, List<OrderHint> orderHints, List<String> propertyPaths);
-	
+
 	/**
 	 * Returns a count of Annotations, optionally filtered by creator and status
-	 * 
+	 *
 	 * @param commentator The user that created the annotation (null to count all annotations, regardless of who made the comment)
 	 * @param status The status of the annotations (null to count all annotations regardless of status)
 	 * @return a count of Annotation instances
 	 */
-    public int count(User creator, MarkerType status);
-	
+    public long count(User creator, MarkerType status);
+
     /**
 	 * Returns a List of Annotations, optionally filtered by creator and status
-	 * 
+	 *
 	 * @param commentator The user that created the annotation (null to list all annotations, regardless of who made the comment)
 	 * @param status The status of the annotations (null to return annotations regardless of status)
 	 * @param pageSize The maximum number of annotations returned (can be null for all annotations)
